nba_dash/requirements.txt
python-dotenv
mysql-connector-python
pandas
numpy
nba_api
pymysql
config